EPMRPP-57964 - React. Tooltip with project name is displayed under the list of projects
