+++ app/src/layouts/common/projectSelector/index.js
@@ -15,4 +15,3 @@
  */
 
 export { ProjectSelector } from './projectSelector';
-export { ProjectSelectorWithTooltip } from './projectSelectorWithTooltip';
