--- app/src/layouts/common/projectSelector/projectSelectorTooltip.scss
@@ -1,18 +1,0 @@
-.tooltip-trigger {
-  display: block;
-  height: 100%;
-}
-
-.project-selector-tooltip {
-  display: inline-block;
-  max-width: 200px;
-  height: 25px;
-  box-sizing: border-box;
-  padding-top: 5px;
-  font-size: 17px;
-  font-family: $FONT-SEMIBOLD;
-  text-transform: uppercase;
-  white-space: pre;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
